+++ test/test_array.cpp
@@ -1598,7 +1598,7 @@ TEST(Array_cares_about)
         0x7fffffffffffff,   0xffffffffffffff,   0x1ffffffffffffff,  0x3ffffffffffffff,  0x7ffffffffffffff,
         0xfffffffffffffff,  0x1fffffffffffffff, 0x3fffffffffffffff, 0x7fffffffffffffff, 0xffffffffffffffff};
     std::vector<uint64_t> res;
-    for (size_t i = 0; i <= 64; i++) {
+    for (uint8_t i = 0; i <= 64; i++) {
         res.push_back(cares_about(i));
     }
     CHECK_EQUAL(res, expected);
@@ -1818,7 +1818,7 @@ TEST(ParallelSearchEqualMatch)
     constexpr size_t buflen = 4;
     uint64_t buff[buflen];
     std::vector<int64_t> values;
-    for (size_t width = 1; width <= 64; width++) {
+    for (uint8_t width = 1; width <= 64; width++) {
         const size_t size = (buflen * 64) / width;
         const uint64_t bit_mask = 0xFFFFFFFFFFFFFFFFULL >> (64 - width); // (1ULL << width) - 1;
